requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alibabacloud-darabonba-string==0.0.4
alibabacloud-ecs20140526==2.1.0
alibabacloud-tea==0.3.5
alibabacloud_credentials==0.3.2
alibabacloud_darabonba_env==0.0.1
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.1
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_console==0.0.1
alibabacloud_tea_openapi==0.3.8
alibabacloud_tea_util==0.3.11
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-ecs==4.24.71
ansible==9.4.0
ansible-core==2.16.5
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-rest-swagger==2.2.0
djangorestframework==3.15.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
jmespath==0.10.0
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openapi-codec==1.3.2
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
resolvelib==1.0.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tencentcloud-sdk-python==3.0.1114
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlrd==1.2.0
yarl==1.9.4
